añado api_tests y correcciones meneros del codigo

+++ views/comentarios.py
@@ -2,7 +2,7 @@ from flask.views import MethodView
 from flask import jsonify, request
 from services.comentario_service import ComentarioService
 from schemas import ComentarioSchema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 
 comentario_schema = ComentarioSchema()
 comentarios_schema = ComentarioSchema(many=True)
@@ -17,9 +17,12 @@ class ComentarioAPI(MethodView):
     @jwt_required()
     def post(self, post_id):
         data = request.get_json()
+        current_user_id = get_jwt_identity() # Obtenemo el id del token
+        
         texto = data.get("texto")
-        usuario_id = data.get("usuario_id")
-        nuevo_comentario = ComentarioService.crear_comentario(texto, usuario_id, post_id)
+        
+        #usa el id del token
+        nuevo_comentario = ComentarioService.crear_comentario(texto, current_user_id, post_id)
         return jsonify(comentario_schema.dump(nuevo_comentario)), 201
 
     @jwt_required()
@@ -27,5 +30,15 @@ class ComentarioAPI(MethodView):
         comentario = ComentarioService.repo.get_by_id(comment_id)
         if not comentario:
             return jsonify({"msg": "Comentario no encontrado"}), 404
+            
+        claims = get_jwt()
+        current_user_id = claims.get("sub")
+        current_user_role = claims.get("role")
+        
+        # Permite al admin, al moderador, o al autor del comentario
+        allowed = ["admin", "moderator"]
+        if current_user_role not in allowed and comentario.usuario_id != current_user_id:
+            return jsonify({"msg": "Permiso denegado"}), 403
+            
         ComentarioService.eliminar_comentario(comentario)
         return jsonify({"msg": "Comentario eliminado"})
